requirements.txt
Django>=6.0
psycopg2-binary
gunicorn
pillow
django-crispy-forms
django-ckeditor
whitenoise
dj-database-url